@@ -5,7 +5,7 @@ import { AppRoutingModule } from './app-routing.module';
 import { provideAnimationsAsync } from '@angular/platform-browser/animations/async'
 import { MatToolbar } from "@angular/material/toolbar";
 import { MatAnchor } from "@angular/material/button";
-import { HttpClient, HttpClientModule, provideHttpClient, withInterceptors } from "@angular/common/http";
+import { HttpClient, provideHttpClient, withInterceptors, withInterceptorsFromDi } from "@angular/common/http";
 import { TranslateLoader, TranslateModule } from "@ngx-translate/core";
 import { TranslateHttpLoader } from "@ngx-translate/http-loader";
 import { MatButtonToggleModule } from "@angular/material/button-toggle";
@@ -63,87 +63,80 @@ export function HttpLoaderFactory(http: HttpClient) {
   return new TranslateHttpLoader(http, './assets/i18n/', '.json');
 }
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    PageNotFoundComponent,
-    LanguageSwitcherComponent,
-    PageCreateGroupComponent,
-    FormCreateGroupComponent,
-    HomeComponent,
-    PageGroupComponent,
-    IncomingComponent,
-    OutgoingComponent,
-    ContactComponent,
-    FormCreateContactComponent,
-    TransactionsTimelineComponent,
-    ChartComponent,
-    PageExpensesComponent,
-    ExpenseCardComponent,
-    AddExpenseComponent,
-    AddPaymentComponent,
-    FormExpenseComponent,
-    FormPaymentComponent,
-    PageGroupDetailsComponent,
-    AuthenticationSectionComponent,
-    SignInComponent,
-    SignUpComponent,],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    MatToolbar,
-    MatAnchor,
-    HeaderComponent,
-    HttpClientModule,
-    MatProgressSpinnerModule,
-    MatButtonToggleModule,
-    TranslateModule.forRoot({
-      defaultLanguage: 'en',
-      loader: {
-        provide: TranslateLoader,
-        useFactory: HttpLoaderFactory,
-        deps: [HttpClient]
-      }
-    }),
-    MatInputModule,
-    MatDatepickerModule,
-    MatNativeDateModule,
-    MatFormFieldModule,
-    MatStepperModule,
-    MatButtonModule,
-    FormsModule,
-    ReactiveFormsModule,
-    MatIcon,
-    MatSidenavContent,
-    MatListItemAvatar,
-    MatListItemLine,
-    MatSidenavContainer,
-    MatNavList,
-    MatListItem,
-    MatSidenav,
-    MatCard,
-    MatCardModule,
-    MatCardHeader,
-    MatCardTitleGroup,
-    MatDialogModule,
-    MatIconModule,
-    MatCheckbox,
-    MatRadioButton,
-    ChartModule,
-    MatSelectModule,
-    MatAutocompleteModule,
-    MatButtonModule,
-    MatIconModule,
-    MatSlideToggle,
-  ],
-  providers: [
-    provideAnimationsAsync(),
-    provideHttpClient(withInterceptors([authenticationInterceptor])),
-    GroupService,
-  ],
-  exports: [
-
-  ],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        PageNotFoundComponent,
+        LanguageSwitcherComponent,
+        PageCreateGroupComponent,
+        FormCreateGroupComponent,
+        HomeComponent,
+        PageGroupComponent,
+        IncomingComponent,
+        OutgoingComponent,
+        ContactComponent,
+        FormCreateContactComponent,
+        TransactionsTimelineComponent,
+        ChartComponent,
+        PageExpensesComponent,
+        ExpenseCardComponent,
+        AddExpenseComponent,
+        AddPaymentComponent,
+        FormExpenseComponent,
+        FormPaymentComponent,
+        PageGroupDetailsComponent,
+        AuthenticationSectionComponent,
+        SignInComponent,
+        SignUpComponent,
+    ],
+    exports: [],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        AppRoutingModule,
+        MatToolbar,
+        MatAnchor,
+        HeaderComponent,
+        MatProgressSpinnerModule,
+        MatButtonToggleModule,
+        TranslateModule.forRoot({
+            defaultLanguage: 'en',
+            loader: {
+                provide: TranslateLoader,
+                useFactory: HttpLoaderFactory,
+                deps: [HttpClient]
+            }
+        }),
+        MatInputModule,
+        MatDatepickerModule,
+        MatNativeDateModule,
+        MatFormFieldModule,
+        MatStepperModule,
+        MatButtonModule,
+        FormsModule,
+        ReactiveFormsModule,
+        MatIcon,
+        MatSidenavContent,
+        MatListItemAvatar,
+        MatListItemLine,
+        MatSidenavContainer,
+        MatNavList,
+        MatListItem,
+        MatSidenav,
+        MatCard,
+        MatCardModule,
+        MatCardHeader,
+        MatCardTitleGroup,
+        MatDialogModule,
+        MatIconModule,
+        MatCheckbox,
+        MatRadioButton,
+        ChartModule,
+        MatSelectModule,
+        MatAutocompleteModule,
+        MatButtonModule,
+        MatIconModule,
+        MatSlideToggle], providers: [
+        provideAnimationsAsync(),
+        provideHttpClient(withInterceptors([authenticationInterceptor])),
+        GroupService,
+        provideHttpClient(withInterceptorsFromDi()),
+    ] })
 export class AppModule { }
